// src/types.ts
export type ColorPalette = {
  bg0: string;
  bg1: string;
  bg: string;
  bg2: string;
  bg3: string;
  bg4: string;
  bg5: string;
  bg6: string;
  bg7: string;
  bg8: string;
  bg9: string;
  grey0: string;
  grey1: string;
  grey2: string;
  fg0: string;
  fg: string;
  fg1: string;
  accent1: string;
  accent2: string;
  accent3: string;
  accent4: string;
  accent5: string;
  accent6: string;
  accent7: string;
  dimAccent1: string;
  dimAccent2: string;
  dimAccent3: string;
  dimAccent4: string;
  dimAccent5: string;
  dimAccent6: string;
  dimAccent7: string;
  selGreen: string;
  selRed: string;
  selBlue: string;
  gitDiffGreen: string;
  gitDiffRed: string;
  gitDiffBlue: string;
  shadow: string;
};